Use `precmd` to print multi-line prompt.
Multiline prompts run into issues with some versions of zsh, so print
first two prompt lines with `precmd`. `precmd` does not appear to have
access to the return status of the previous command, so this approach
doesn't display a red return line when the command exits with a non-zero
status.